Ubuntu 19.04 scripts need an explicit TERM environment variable.
Without a TTY, and thus TERM value, any 19.04 script which called
"tput" was ending with a fatal error. As such we now detect the
missing TERM value and explicitly set the value to "vt100" to avoid
spurrious errors. The actual changes were accidentally added to the
index as part of the 651e8bd1a0846b2340d352332ef5a6380e4f97ae commit.