Change storage interface for RHEL8
By default, the virtualbox-iso builder will use an IDE controller for
the boot drive as well as the CD drive. The performance of the emulated
SATA Controller is presumed to be better and IDE support may be phased
out in future kernel versions. This uses a SATA controller with 6
virtual ports instead, allowing you to attach additional drives to your
vagrant box.